backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.109.0              # Основной веб-фреймворк
uvicorn[standard]==0.27.0     # ASGI сервер для FastAPI
python-multipart==0.0.6       # Обработка multipart/form-data (загрузка файлов)
aiofiles==23.2.1              # Асинхронная работа с файлами
httpx>=0.25.0                 # HTTP клиент для внешних запросов

# Обработка изображений
pillow>=10.0.0                # Базовая работа с изображениями (PNG, JPEG, GIF, BMP, TIFF)
opencv-python>=4.8.0          # Компьютерное зрение, перспективные трансформации
numpy>=1.24.0                 # Работа с массивами изображений
psd-tools>=1.9.0              # Поддержка PSD файлов (Photoshop)

# Валидация и модели данных
pydantic>=2.0.0               # Валидация данных и сериализация

# Машинное обучение (для анализа цветов/кластеризации)
scikit-learn>=1.3.0           # KMeans для определения доминантных цветов

# Работа с Excel
openpyxl>=3.1.0               # Импорт/экспорт Excel файлов

# Compression
starlette>=0.35.0             # GZip middleware (included with FastAPI)
